Separado os componentes da mainActivity
