+++ app/src/main/java/com/example/aluvery/MainActivity.kt
@@ -3,46 +3,10 @@ package com.example.aluvery
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.horizontalScroll
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment.Companion.BottomCenter
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.example.aluvery.extensions.toBrazilianCurrency
-import com.example.aluvery.model.ProductModel
+import com.example.aluvery.ui.screens.HomeScreen
 import com.example.aluvery.ui.theme.AluveryTheme
-import com.example.aluvery.ui.theme.Purple500
-import com.example.aluvery.ui.theme.Teal200
-import java.math.BigDecimal
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -57,148 +21,9 @@ class MainActivity : ComponentActivity() {
 fun App(){
     AluveryTheme {
         Surface {
-            Column(
-                Modifier
-                    .fillMaxSize()
-                    .verticalScroll(rememberScrollState()),
-                verticalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                Spacer(modifier = Modifier)
-                ProductSection("Promoções")
-                ProductSection("Doces")
-                ProductSection("Bebidas")
-                Spacer(modifier = Modifier)
-            }
+            HomeScreen()
         }
     }
 
 
-}
-
-@Composable
-fun ProductItem(productModel: ProductModel) {
-    Surface(
-        shape = RoundedCornerShape(15.dp),
-        shadowElevation = 8.dp
-    ) {
-        Column(
-            Modifier
-                .heightIn(250.dp, 300.dp)
-                .width(200.dp),
-
-            ) {
-            val imageSize = 100.dp
-            Box(
-                modifier = Modifier
-                    .height(imageSize)
-                    .background(
-                        brush = Brush.horizontalGradient(
-                            colors = listOf(
-                                Purple500, Teal200
-                            )
-                        )
-                    )
-                    .fillMaxWidth()
-            ) {
-                Image(
-                    painter = painterResource(id = productModel.image),
-                    contentDescription = null,
-                    Modifier
-                        .size(imageSize)
-                        .offset(y = imageSize / 2)
-                        .clip(shape = CircleShape)
-                        .align(BottomCenter),
-                    contentScale = ContentScale.Crop
-
-                )
-            }
-            Spacer(modifier = Modifier.height(imageSize / 2))
-            Column(Modifier.padding(16.dp)) {
-                Text(
-                    text = productModel.name,
-                    fontSize = 18.sp,
-                    fontWeight = FontWeight(700),
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis
-
-                )
-                Text(
-                    text = productModel.value.toBrazilianCurrency(),
-                    Modifier.padding(top = 8.dp),
-                    fontSize = 14.sp,
-                    fontWeight = FontWeight(400)
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun ProductSection(topic: String) {
-    Column {
-        Text(
-            text = topic,
-            Modifier.padding(
-                start = 16.dp
-            ),
-            fontSize = 20.sp,
-            fontWeight = FontWeight(400)
-        )
-        Row(
-            Modifier
-                .padding(
-                    top = 8.dp
-                )
-                .fillMaxWidth()
-                .horizontalScroll(rememberScrollState()),
-            horizontalArrangement = Arrangement.spacedBy(16.dp)
-        ) {
-            Spacer(Modifier)
-            ProductItem(
-                ProductModel(
-                    name = "meia Tietê com borda brumadinho",
-                    value = BigDecimal("39.99"),
-                    image = R.drawable.pizza
-                )
-            )
-            ProductItem(
-                ProductModel(
-                    name = "x-ratão",
-                    value = BigDecimal("10.99"),
-                    image = R.drawable.burger
-                )
-            )
-            ProductItem(
-                ProductModel(
-                    name = "porção batata 200gr",
-                    value = BigDecimal("5.90"),
-                    image = R.drawable.fries
-                )
-            )
-            Spacer(Modifier)
-        }
-    }
-}
-
-@Preview(showSystemUi = true)
-@Composable
-fun AppPreview(){
-    App()
-}
-@Preview(showBackground = true)
-@Composable
-fun ProductItemPreview() {
-    ProductItem(
-        ProductModel(
-            name = "lorem",
-            value = BigDecimal("14.99"),
-            image = R.drawable.burger
-        )
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun ProductSectionPreview() {
-    ProductSection("Promoções")
 }
