+++ app/src/main/java/com/example/aluvery/ui/components/ProductItem.kt
@@ -1,0 +1,106 @@
+package com.example.aluvery.ui.components
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.heightIn
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.aluvery.R
+import com.example.aluvery.extensions.toBrazilianCurrency
+import com.example.aluvery.model.ProductModel
+import com.example.aluvery.ui.theme.Purple500
+import com.example.aluvery.ui.theme.Teal200
+import java.math.BigDecimal
+
+@Composable
+fun ProductItem(productModel: ProductModel) {
+    Surface(
+        shape = RoundedCornerShape(15.dp),
+        shadowElevation = 8.dp
+    ) {
+        Column(
+            Modifier
+                .heightIn(250.dp, 300.dp)
+                .width(200.dp),
+
+            ) {
+            val imageSize = 100.dp
+            Box(
+                modifier = Modifier
+                    .height(imageSize)
+                    .background(
+                        brush = Brush.horizontalGradient(
+                            colors = listOf(
+                                Purple500, Teal200
+                            )
+                        )
+                    )
+                    .fillMaxWidth()
+            ) {
+                Image(
+                    painter = painterResource(id = productModel.image),
+                    contentDescription = null,
+                    Modifier
+                        .size(imageSize)
+                        .offset(y = imageSize / 2)
+                        .clip(shape = CircleShape)
+                        .align(Alignment.BottomCenter),
+                    contentScale = ContentScale.Crop
+
+                )
+            }
+            Spacer(modifier = Modifier.height(imageSize / 2))
+            Column(Modifier.padding(16.dp)) {
+                Text(
+                    text = productModel.name,
+                    fontSize = 18.sp,
+                    fontWeight = FontWeight(700),
+                    maxLines = 2,
+                    overflow = TextOverflow.Ellipsis
+
+                )
+                Text(
+                    text = productModel.value.toBrazilianCurrency(),
+                    Modifier.padding(top = 8.dp),
+                    fontSize = 14.sp,
+                    fontWeight = FontWeight(400)
+                )
+            }
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun ProductItemPreview() {
+    ProductItem(
+        ProductModel(
+            name = "lorem",
+            value = BigDecimal("14.99"),
+            image = R.drawable.placeholder
+        )
+    )
+}
